+++ src/js/3D/exporters/M2Exporter.js
@@ -22,11 +22,11 @@ class M2Exporter {
 	/**
 	 * Construct a new M2Exporter instance.
 	 * @param {BufferWrapper}
-	 * @param {number} variantTexture
+	 * @param {Array} variantTextures
 	 */
-	constructor(data, variantTexture = 0) {
+	constructor(data, variantTextures) {
 		this.m2 = new M2Loader(data);
-		this.variantTexture = variantTexture;
+		this.variantTextures = variantTextures;
 	}
 
 	/**
@@ -53,20 +53,30 @@ class M2Exporter {
 		const useAlpha = config.modelsIncludeAlpha;
 
 		const validTextures = new Map();
+		let textureIndex = 0;
 		for (const texture of this.m2.textures) {
 			// Abort if the export has been cancelled.
 			if (helper.isCancelled())
 				return;
-				
+
+			const textureType = this.m2.textureTypes[textureIndex];
 			let texFileDataID = texture.fileDataID;
 
-			// Blank texture, do we have a variant texture?
-			if (texFileDataID === 0) {
-				texFileDataID = this.variantTexture;
+			if (textureType > 0) {
+				let targetFileDataID = 0;
+
+				if (textureType >= 11 && textureType < 14) {
+					// Creature textures.
+					targetFileDataID = this.variantTextures[textureType - 11];
+				} else if (textureType > 1 && textureType < 5) {
+					targetFileDataID = this.variantTextures[textureType - 2];
+				}
+
+				texFileDataID = targetFileDataID;
 
 				// Backward patch the variant texture into the M2 instance so that
 				// the MTL exports with the correct texture once we swap it here.
-				texture.fileDataID = this.variantTexture;
+				texture.fileDataID = targetFileDataID;
 			}
 
 			if (texFileDataID > 0) {
@@ -123,6 +133,8 @@ class M2Exporter {
 					log.write('Failed to export texture %d for M2: %s', texFileDataID, e.message);
 				}
 			}
+
+			textureIndex++;
 		}
 
 		return validTextures;