+++ src/js/casc/casc-source-local.js
@@ -91,7 +91,7 @@ class CASCLocal extends CASC {
 		this.cache = new BuildCache(this.build.BuildKey);
 		await this.cache.init();
 
-		this.progress = core.createProgress(8);
+		this.progress = core.createProgress(12);
 		await this.loadConfigs();
 		await this.loadIndexes();
 		await this.loadEncoding();
@@ -100,6 +100,7 @@ class CASCLocal extends CASC {
 		core.view.casc = this;
 
 		await this.loadListfile(this.build.BuildKey);
+		await this.loadTables();
 		await this.initializeComponents();
 	}
 
@@ -334,6 +335,14 @@ class CASCLocal extends CASC {
 	*/
 	getBuildName() {
 		return this.build.Version;
+	}
+
+	/**
+	 * Returns the build configuration key.
+	 * @returns {string}
+	 */
+	getBuildKey() {
+		return this.build.BuildKey;
 	}
 }