+++ src/js/casc/casc-source.js
@@ -1,6 +1,6 @@
 /*!
 	wow.export (https://github.com/Kruithne/wow.export)
-	Authors: Kruithne <kruithne@gmail.com>
+	Authors: Kruithne <kruithne@gmail.com>, Martin Benjamins <marlamin@marlamin.com>
 	License: MIT
  */
 const BufferWrapper = require('../buffer');
@@ -11,6 +11,12 @@ const core = require('../core');
 const constants = require('../constants');
 const LocaleFlag = require('./locale-flags').flags;
 const ContentFlag = require('./content-flags');
+
+const WDCReader = require('../db/WDCReader');
+const DBModelFileData = require('../db/caches/DBModelFileData');
+const DBTextureFileData = require('../db/caches/DBTextureFileData');
+const DBItemDisplays = require('../db/caches/DBItemDisplays');
+const DBCreatures = require('../db/caches/DBCreatures');
 
 const ENC_MAGIC = 0x4E45;
 const ROOT_MAGIC = 0x4D465354;
@@ -107,6 +113,33 @@ class CASC {
 		core.view.listfileVideos = listfile.getFilenamesByExtension('.avi');
 		core.view.listfileText = listfile.getFilenamesByExtension(['.txt', '.lua', '.xml', '.sbt', '.wtf', '.htm', '.toc', '.xsd']);
 		core.view.listfileModels = listfile.getFilenamesByExtension(modelExt);
+	}
+
+	/**
+	 * Load tables that are required globally.
+	 */
+	async loadTables() {
+		await this.progress.step('Loading model file data');
+		await DBModelFileData.initializeModelFileData();
+
+		await this.progress.step('Loading texture file data');
+		await DBTextureFileData.initializeTextureFileData();
+
+		// Once the above two tables have loaded, ingest fileDataIDs as
+		// unknown entries to the listfile.
+		await listfile.loadUnknowns();
+
+		await this.progress.step('Loading item displays');
+		await DBItemDisplays.initializeItemDisplays();
+
+		await this.progress.step('Loading creature data');
+		const creatureDisplayInfo = new WDCReader('DBFilesClient/CreatureDisplayInfo.db2');
+		await creatureDisplayInfo.parse();
+
+		const creatureModelData = new WDCReader('DBFilesClient/CreatureModelData.db2');
+		await creatureModelData.parse();
+
+		await DBCreatures.initializeCreatureData(creatureDisplayInfo, creatureModelData);
 	}
 
 	/**