+++ src/js/casc/listfile.js
@@ -12,6 +12,8 @@ const BufferWrapper = require('../buffer');
 const ExportHelper = require('../casc/export-helper');
 
 const WDCReader = require('../db/WDCReader');
+const DBTextureFileData = require('../db/caches/DBTextureFileData');
+const DBModelFileData = require('../db/caches/DBModelFileData');
 
 const nameLookup = new Map();
 const idLookup = new Map();
@@ -110,26 +112,31 @@ const loadListfile = async (buildConfig, cache) => {
 		nameLookup.set(fileName, fileDataID);
 	}
 
-	let unknownCount = 0;
-	unknownCount += await loadIDTable('DBFilesClient/ModelFileData.db2', '.m2');
-	unknownCount += await loadIDTable('DBFilesClient/TextureFileData.db2', '.blp');
-
-	log.write('%d listfile entries loaded (%d unknown entries)', idLookup.size, unknownCount);
+	log.write('%d listfile entries loaded', idLookup.size);
 	return idLookup.size;
 }
 
 /**
+ * Load unknown files from TextureFileData/ModelFileData.
+ * Must be called after DBTextureFileData/DBModelFileData have loaded.
+ */
+const loadUnknowns = async () => {
+	let unknownCount = 0;
+	unknownCount += await loadIDTable(DBTextureFileData.getFileDataIDs(), '.blp');
+	unknownCount += await loadIDTable(DBModelFileData.getFileDataIDs(), '.m2');
+
+	log.write('Added %d unknown entries to listfile', unknownCount);
+};
+
+/**
  * Load file IDs from a data table.
- * @param {string} tableFile 
+ * @param {Set} ids
  * @param {string} ext 
  */
-const loadIDTable = async (tableFile, ext) => {
+const loadIDTable = async (ids, ext) => {
 	let loadCount = 0;
-	const table = new WDCReader(tableFile);
-	await table.parse();
 
-	for (const row of table.rows.values()) {
-		const fileDataID = row.FileDataID;
+	for (const fileDataID of ids) {
 		if (!idLookup.has(fileDataID)) {
 			const fileName = 'unknown_' + fileDataID + ext;
 			idLookup.set(fileDataID, fileName);
@@ -137,6 +144,10 @@ const loadIDTable = async (tableFile, ext) => {
 			loadCount++;
 		}
 	}
+
+	// Reset the fileDataID caches for each table to conserve memory.
+	// If these sets are needed elsewhere, remove this call.
+	ids.clear();
 
 	return loadCount;
 };
@@ -196,4 +207,10 @@ const getByFilename = (filename) => {
 	return lookup;
 };
 
-module.exports = { loadListfile, getByID, getByFilename, getFilenamesByExtension };
+module.exports = {
+	loadListfile,
+	loadUnknowns,
+	getByID,
+	getByFilename,
+	getFilenamesByExtension
+};